+++ pkg/cmd/root.go
@@ -52,9 +52,10 @@ func NewRootCmd(
 
 	// add subcommands
 	rootCmd.AddCommand(NewPodCmd(duplicator, client))
-	rootCmd.AddCommand(NewListDuplicatedCmd(client))
 	rootCmd.AddCommand(NewDeployCmd(duplicator, client))
 	rootCmd.AddCommand(NewStatefulSetCmd(duplicator, client))
+	rootCmd.AddCommand(NewListDuplicatedCmd(client))
+	rootCmd.AddCommand(NewCleanupCmd(client))
 
 	return rootCmd
 }