+++ pkg/cmd/shared.go
@@ -19,6 +19,8 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/charmbracelet/lipgloss/table"
 	"github.com/spf13/cobra"
 	"github.com/telemaco019/duplik8s/pkg/clients"
 	"github.com/telemaco019/duplik8s/pkg/cmd/flags"
@@ -108,4 +110,28 @@ func addOverrideFlags(cmd *cobra.Command) {
 		[]string{"-c", "trap 'exit 0' INT TERM KILL; while true; do sleep 1; done"},
 		"Override the command of each container in the Pod.",
 	)
+}
+
+func renderDuplicatedObjects(duplicatedObjs []core.DuplicatedObject) {
+	headerStyle := lipgloss.NewStyle().Bold(true).Padding(0, 1)
+	defaultStyle := lipgloss.NewStyle().Padding(0, 1)
+	t := table.New().Border(lipgloss.HiddenBorder()).
+		StyleFunc(func(row, col int) lipgloss.Style {
+			switch {
+			case row == 0:
+				return headerStyle
+			default:
+				return defaultStyle
+			}
+		}).
+		Headers("Namespace", "Kind", "Name", "Age")
+	for _, obj := range duplicatedObjs {
+		t.Row(
+			obj.Namespace,
+			obj.ObjectKind.GroupVersionKind().Kind,
+			obj.Name,
+			utils.FormatAge(obj.CreationTimestamp),
+		)
+	}
+	fmt.Print(t.Render() + "\n")
 }